Splitted out the refactoring of entity dialog from the download cancel changes
