+++ base-ui/api/current.api
@@ -1,9 +1,9 @@
 // Signature format: 4.0
 package com.google.android.horologist.base.ui.components {
 
-  public final class ConfirmationDialogKt {
-    method @androidx.compose.runtime.Composable public static void ConfirmationDialog(String prompt, String proceedText, String cancelText, kotlin.jvm.functions.Function0<kotlin.Unit> onCancelButtonClick, kotlin.jvm.functions.Function0<kotlin.Unit> onProceedButtonClick, boolean showDialog, androidx.wear.compose.material.ScalingLazyListState scalingLazyListState);
-    method @androidx.compose.runtime.Composable public static void ConfirmationDialogAlert(String prompt, String proceedText, String cancelText, kotlin.jvm.functions.Function0<kotlin.Unit> onCancelButtonClick, kotlin.jvm.functions.Function0<kotlin.Unit> onProceedButtonClick);
+  public final class AlertDialogKt {
+    method @androidx.compose.runtime.Composable public static void AlertDialog(String text, String proceedText, String cancelText, kotlin.jvm.functions.Function0<kotlin.Unit> onCancelButtonClick, kotlin.jvm.functions.Function0<kotlin.Unit> onProceedButtonClick, boolean showDialog, androidx.wear.compose.material.ScalingLazyListState scalingLazyListState);
+    method @androidx.compose.runtime.Composable public static void AlertDialogAlert(String text, String proceedText, String cancelText, kotlin.jvm.functions.Function0<kotlin.Unit> onCancelButtonClick, kotlin.jvm.functions.Function0<kotlin.Unit> onProceedButtonClick);
   }
 
   public final class StandardButtonKt {
