+++ media-sample/src/main/java/com/google/android/horologist/mediasample/di/DataModule.kt
@@ -32,10 +32,8 @@ import com.google.android.horologist.media.data.mapper.MediaExtrasMapperNoopImpl
 import com.google.android.horologist.media.data.mapper.MediaMapper
 import com.google.android.horologist.media.data.mapper.PlaylistDownloadMapper
 import com.google.android.horologist.media.data.mapper.PlaylistMapper
-import com.google.android.horologist.media.data.repository.MediaDownloadRepositoryImpl
 import com.google.android.horologist.media.data.repository.PlaylistDownloadRepositoryImpl
 import com.google.android.horologist.media.data.repository.PlaylistRepositoryImpl
-import com.google.android.horologist.media.repository.MediaDownloadRepository
 import com.google.android.horologist.media.repository.PlaylistDownloadRepository
 import com.google.android.horologist.media.repository.PlaylistRepository
 import com.google.android.horologist.mediasample.data.api.NetworkChangeListService
@@ -75,15 +73,6 @@ class DataModule {
             mediaDownloadLocalDataSource = mediaDownloadLocalDataSource,
             media3DownloadDataSource = media3DownloadDataSource,
             playlistDownloadMapper = playlistDownloadMapper
-        )
-
-    @Singleton
-    @Provides
-    fun mediaDownloadRepository(
-        media3DownloadDataSource: Media3DownloadDataSource
-    ): MediaDownloadRepository =
-        MediaDownloadRepositoryImpl(
-            media3DownloadDataSource = media3DownloadDataSource
         )
 
     @Singleton
